refactor(weather): Move weather widgets to UI layer
- Relocated weather-related widgets from features/weather/components/ to ui/weather/widgets/
- Updated all internal imports to reflect new paths
- Promotes better separation of concerns between UI and domain logic
